# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Sentinel Authorized Pentest Platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Scope {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  signedBy: string;
  roeActive: boolean;
  /** @nullable */
  validFrom?: string | null;
  /** @nullable */
  validUntil?: string | null;
  createdAt: string;
  targetCount?: number;
}

export interface ScopeInput {
  /** @minLength 1 */
  name: string;
  description?: string;
  /** @minLength 1 */
  signedBy: string;
  roeActive?: boolean;
  validFrom?: string;
  validUntil?: string;
}

export interface ScopeUpdate {
  name?: string;
  description?: string;
  signedBy?: string;
  roeActive?: boolean;
  validFrom?: string;
  validUntil?: string;
}

export interface Target {
  id: number;
  scopeId: number;
  scopeName?: string;
  host: string;
  /** @nullable */
  ip?: string | null;
  /** @nullable */
  portRange?: string | null;
  allowed: boolean;
  activeModeEnabled: boolean;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
}

export interface TargetInput {
  scopeId: number;
  /** @minLength 1 */
  host: string;
  ip?: string;
  portRange?: string;
  allowed?: boolean;
  activeModeEnabled?: boolean;
  notes?: string;
}

export interface TargetUpdate {
  host?: string;
  ip?: string;
  portRange?: string;
  allowed?: boolean;
  activeModeEnabled?: boolean;
  notes?: string;
}

export type ScanPhase = typeof ScanPhase[keyof typeof ScanPhase];


export const ScanPhase = {
  recon: 'recon',
  scan: 'scan',
  exploit: 'exploit',
  intel: 'intel',
} as const;

export type ScanStatus = typeof ScanStatus[keyof typeof ScanStatus];


export const ScanStatus = {
  queued: 'queued',
  running: 'running',
  completed: 'completed',
  failed: 'failed',
  cancelled: 'cancelled',
} as const;

export interface Scan {
  id: number;
  targetId: number;
  targetHost?: string;
  scopeName?: string;
  tool: string;
  phase: ScanPhase;
  status: ScanStatus;
  /** @nullable */
  startedAt?: string | null;
  /** @nullable */
  completedAt?: string | null;
  /** @nullable */
  errorMessage?: string | null;
  findingCount?: number;
  createdAt: string;
}

export type ScanInputPhase = typeof ScanInputPhase[keyof typeof ScanInputPhase];


export const ScanInputPhase = {
  recon: 'recon',
  scan: 'scan',
  exploit: 'exploit',
  intel: 'intel',
} as const;

export type ScanInputOptions = { [key: string]: unknown };

export interface ScanInput {
  targetId: number;
  /** @minLength 1 */
  tool: string;
  phase: ScanInputPhase;
  options?: ScanInputOptions;
}

export type FindingSeverity = typeof FindingSeverity[keyof typeof FindingSeverity];


export const FindingSeverity = {
  critical: 'critical',
  high: 'high',
  medium: 'medium',
  low: 'low',
  info: 'info',
} as const;

export interface Finding {
  id: number;
  scanId: number;
  targetId: number;
  targetHost?: string;
  tool: string;
  severity: FindingSeverity;
  title: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  evidence?: string | null;
  cveRefs?: string[];
  /** @nullable */
  remediation?: string | null;
  /** @nullable */
  raw?: string | null;
  createdAt: string;
}

export interface SeveritySummary {
  critical: number;
  high: number;
  medium: number;
  low: number;
  info: number;
}

export type AuditEntryOutcome = typeof AuditEntryOutcome[keyof typeof AuditEntryOutcome];


export const AuditEntryOutcome = {
  allowed: 'allowed',
  denied: 'denied',
} as const;

export interface AuditEntry {
  id: number;
  operator: string;
  /** @nullable */
  targetId?: number | null;
  targetHost: string;
  tool: string;
  /** @nullable */
  flags?: string | null;
  /** @nullable */
  authorizedBy?: string | null;
  outcome: AuditEntryOutcome;
  createdAt: string;
}

export type ReportFormat = typeof ReportFormat[keyof typeof ReportFormat];


export const ReportFormat = {
  json: 'json',
  html: 'html',
} as const;

export interface Report {
  id: number;
  scanId: number;
  targetHost?: string;
  format: ReportFormat;
  /** @nullable */
  summary?: string | null;
  findingCount?: number;
  generatedAt: string;
}

export type ReportInputFormat = typeof ReportInputFormat[keyof typeof ReportInputFormat];


export const ReportInputFormat = {
  json: 'json',
  html: 'html',
} as const;

export interface ReportInput {
  scanId: number;
  format: ReportInputFormat;
}

export interface DailyCount {
  date: string;
  count: number;
}

export interface DashboardSummary {
  totalScopes: number;
  totalTargets: number;
  totalScans: number;
  activeScans: number;
  totalFindings: number;
  criticalFindings: number;
  highFindings: number;
  recentFindingsByDay?: DailyCount[];
}

export type ActivityItemType = typeof ActivityItemType[keyof typeof ActivityItemType];


export const ActivityItemType = {
  scan_started: 'scan_started',
  scan_completed: 'scan_completed',
  scan_failed: 'scan_failed',
  finding_critical: 'finding_critical',
  finding_high: 'finding_high',
  scan_cancelled: 'scan_cancelled',
} as const;

export interface ActivityItem {
  id: string;
  type: ActivityItemType;
  title: string;
  description: string;
  /** @nullable */
  severity?: string | null;
  /** @nullable */
  targetHost?: string | null;
  createdAt: string;
}

export type ListTargetsParams = {
scopeId?: number;
};

export type ListScansParams = {
targetId?: number;
scopeId?: number;
status?: string;
};

export type ListFindingsParams = {
severity?: string;
targetId?: number;
scopeId?: number;
tool?: string;
};

export type ListAuditEntriesParams = {
targetId?: number;
tool?: string;
limit?: number;
};

export type GetRecentActivityParams = {
limit?: number;
};