# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Sentinel Authorized Pentest Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /scans:
    get:
      operationId: listScans
      tags: [scans]
      summary: List scan jobs
      parameters:
        - name: targetId
          in: query
          required: false
          schema:
            type: integer
        - name: scopeId
          in: query
          required: false
          schema:
            type: integer
        - name: status
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of scans
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Scan"
    post:
      operationId: createScan
      tags: [scans]
      summary: Start a new scan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScanInput"
      responses:
        "201":
          description: Created scan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Scan"
        "403":
          description: Target not authorized
  /scans/{id}:
    get:
      operationId: getScan
      tags: [scans]
      summary: Get a scan by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Scan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Scan"
        "404":
          description: Not found
  /scans/{id}/cancel:
    post:
      operationId: cancelScan
      tags: [scans]
      summary: Cancel a running scan
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Scan cancelled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Scan"
  /scans/{id}/findings:
    get:
      operationId: getScanFindings
      tags: [scans]
      summary: Get findings for a scan
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Findings for scan
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Finding"

  # ── Findings ────────────────────────────────────────────────────────────
  /findings:
    get:
      operationId: listFindings
      tags: [findings]
      summary: List all findings
      parameters:
        - name: severity
          in: query
          required: false
          schema:
            type: string
        - name: targetId
          in: query
          required: false
          schema:
            type: integer
        - name: scopeId
          in: query
          required: false
          schema:
            type: integer
        - name: tool
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of findings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Finding"
  /findings/{id}:
    get:
      operationId: getFinding
      tags: [findings]
      summary: Get a finding by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Finding
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Finding"
        "404":
          description: Not found
...
  /reports:
    get:
      operationId: listReports
      tags: [reports]
      summary: List generated reports
      responses:
        "200":
          description: Reports
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Report"
    post:
      operationId: generateReport
      tags: [reports]
      summary: Generate a report for a scan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReportInput"
      responses:
        "201":
          description: Generated report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Report"
  /reports/{id}:
    get:
      operationId: getReport
      tags: [reports]
      summary: Get a report by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Report"
        "404":
          description: Not found

  # ── Dashboard ────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Scan:
      type: object
      required: [id, targetId, tool, phase, status, createdAt]
      properties:
        id:
          type: integer
        targetId:
          type: integer
        targetHost:
          type: string
        scopeName:
          type: string
        tool:
          type: string
        phase:
          type: string
          enum: [recon, scan, exploit, intel]
        status:
          type: string
          enum: [queued, running, completed, failed, cancelled]
        startedAt:
          type: ["string", "null"]
        completedAt:
          type: ["string", "null"]
        errorMessage:
          type: ["string", "null"]
        findingCount:
          type: integer
        createdAt:
          type: string
    ScanInput:
      type: object
      required: [targetId, tool, phase]
      properties:
        targetId:
          type: integer
        tool:
          type: string
          minLength: 1
        phase:
          type: string
          enum: [recon, scan, exploit, intel]
        options:
          type: object
          additionalProperties: true

    # ── Finding ──────────────────────────────────────────────────────────
    Finding:
      type: object
      required: [id, scanId, targetId, tool, severity, title, createdAt]
      properties:
        id:
          type: integer
        scanId:
          type: integer
        targetId:
          type: integer
        targetHost:
          type: string
        tool:
          type: string
        severity:
          type: string
          enum: [critical, high, medium, low, info]
        title:
          type: string
        description:
          type: ["string", "null"]
        evidence:
          type: ["string", "null"]
        cveRefs:
          type: array
          items:
            type: string
        remediation:
          type: ["string", "null"]
        raw:
          type: ["string", "null"]
        createdAt:
          type: string
...
    Report:
      type: object
      required: [id, scanId, format, generatedAt]
      properties:
        id:
          type: integer
        scanId:
          type: integer
        targetHost:
          type: string
        format:
          type: string
          enum: [json, html]
        summary:
          type: ["string", "null"]
        findingCount:
          type: integer
        generatedAt:
          type: string
    ReportInput:
      type: object
      required: [scanId, format]
      properties:
        scanId:
          type: integer
        format:
          type: string
          enum: [json, html]

    # ── Dashboard ─────────────────────────────────────────────────────────
...