# lib/api-zod/src/generated/types/scanInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Sentinel Authorized Pentest Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { ScanInputOptions } from './scanInputOptions';
import type { ScanInputPhase } from './scanInputPhase';

export interface ScanInput {
  targetId: number;
  /** @minLength 1 */
  tool: string;
  phase: ScanInputPhase;
  options?: ScanInputOptions;
}